# docs/openapi/eomp-openapi-spec.yaml
openapi: 3.0.3
info:
  title: EOMP API
  version: 0.1.0
  description: |
    Runtime operation baseline for the EOMP gateway and its downstream services.
    All 102 registered `/api/v1` operations are represented. Generic write bodies
    mark endpoints whose domain schemas still need richer component definitions.
  x-coverage-status: complete-operation-inventory
  x-runtime-operation-count: 102
  x-documented-operation-count: 102
servers:
  - url: http://localhost:8080
    description: Local gateway
security:
  - BearerAuth: []
paths:
...
  /api/v1/approvals:
    get: {tags: [Workflow], operationId: listApprovals, summary: List approvals assigned to the authenticated manager or role pool (admin may list all), responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Forbidden'}}}
...
  /api/v1/assets/employee/{id}/history:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get: {tags: [Assets], operationId: getEmployeeAssetHistory, summary: Get employee asset history, responses: {'200': {$ref: '#/components/responses/Success'}}}
...
  /api/v1/changes:
    get: {tags: [Changes], operationId: listChanges, summary: List change requests (manager or admin), responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Forbidden'}}}
    post: {tags: [Changes], operationId: createChange, summary: Create a change request using the authenticated requester identity, requestBody: {$ref: '#/components/requestBodies/JsonBody'}, responses: {'201': {$ref: '#/components/responses/Created'}, '403': {$ref: '#/components/responses/Forbidden'}}}
  /api/v1/changes/{id}:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get: {tags: [Changes], operationId: getChange, summary: Get a change request, responses: {'200': {$ref: '#/components/responses/Success'}}}
...
  /api/v1/changes/{id}/status:
    parameters: [{$ref: '#/components/parameters/Id'}]
    patch: {tags: [Changes], operationId: updateChangeStatus, summary: Update change status using optimistic locking, requestBody: {$ref: '#/components/requestBodies/VersionedBody'}, responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Conflict'}}}
...
  /api/v1/departments:
    get: {tags: [Employees], operationId: listDepartments, summary: List departments, responses: {'200': {$ref: '#/components/responses/Success'}}}
    post: {tags: [Employees], operationId: createDepartment, summary: Create a department, requestBody: {$ref: '#/components/requestBodies/JsonBody'}, responses: {'201': {$ref: '#/components/responses/Created'}}}
  /api/v1/employees:
    get: {tags: [Employees], operationId: listEmployees, summary: List employees, responses: {'200': {$ref: '#/components/responses/Success'}}}
    post: {tags: [Employees], operationId: createEmployee, summary: Create an employee, requestBody: {$ref: '#/components/requestBodies/JsonBody'}, responses: {'201': {$ref: '#/components/responses/Created'}}}
  /api/v1/employees/{id}:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get: {tags: [Employees], operationId: getEmployee, summary: Get an employee, responses: {'200': {$ref: '#/components/responses/Success'}}}
    put: {tags: [Employees], operationId: updateEmployee, summary: Update an employee using optimistic locking, requestBody: {$ref: '#/components/requestBodies/VersionedBody'}, responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Conflict'}}}
    delete: {tags: [Employees], operationId: deleteEmployee, summary: Delete an employee using optimistic locking, parameters: [{$ref: '#/components/parameters/VersionQuery'}], responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Conflict'}}}
...
  /api/v1/notifications:
    get: {tags: [Notifications], operationId: listNotifications, summary: List notifications, responses: {'200': {$ref: '#/components/responses/Success'}}}
    post: {tags: [Notifications], operationId: sendNotification, summary: Send a notification, requestBody: {$ref: '#/components/requestBodies/JsonBody'}, responses: {'201': {$ref: '#/components/responses/Created'}}}
...
components:
  securitySchemes:
    BearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string}}
...
    VersionQuery: {name: version, in: query, required: true, schema: {type: integer, minimum: 1}, description: Version last read by the caller}
  responses:
    Success: {description: Successful response}
    Created: {description: Resource created}
...
    Forbidden: {description: Authenticated caller lacks permission for this resource}
    Conflict: {description: Optimistic-lock or uniqueness conflict}
  requestBodies:
    JsonBody:
      required: true
      content:
        application/json:
          schema: {type: object, additionalProperties: true}
    VersionedBody:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [version]
            properties:
              version: {type: integer, minimum: 1}
            additionalProperties: true
...